+++ moove/infrastructure/src/main/kotlin/io/charlescd/moove/infrastructure/service/client/response/CreateVillagerRegistryConfigurationResponse.kt
